requirements/base.txt
wheel==0.30.0

Django>=2.0,<2.1

dj-database-url==0.5.0
python-decouple==3.1
